Temporary variable should be an unsigned int.
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@18290 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/other/ftgl/demo/FTGLDemo.cpp
@@ -583,7 +583,7 @@ void parsekey(unsigned char key, int x, int y)
 void parseSpecialKey(int key, int x, int y)
 {
     FTSimpleLayout *l = NULL;
-    float s;
+    unsigned int s;
 
     // If the currentLayout is a SimpleLayout store a pointer in l
     if(layouts[currentLayout]
